Calculate text sentiment indicators using Python packages

Uses Python's TextBlob and vaderSentiment packages to calculate several indicators of sentiment in a text.

calc_sentiment_indicators(
  x,
  python_setup = TRUE,
  sys_setenv,
  which_python,
  which_venv,
  venv_name,
  text_col_name
)

Arguments

x
A data frame with a column that has the text. Any other columns will be ignored.

python_setup
A logical whether to set up the Python version, virtual environment etc. that can be controlled with arguments sys_setenv, which_python, which_venv and venv_name. These arguments will be ignored when python_setup is FALSE. The purpose of python_setup is that users may wish to control the Python parameters outside the actual function, for the session in general.

sys_setenv
A string with the path to the Python version to use in this session.

which_python
A string with the path to the Python version to use in this session.

which_venv
A string that should be "conda", "miniconda" or "python". See use_python{reticulate}.

venv_name
A string with the name of the Python virtual environment to use in this session.

text_col_name
A string with the column name of the text variable.

Value

A data frame with the five indicators described in Details.

Details

TextBlob calculates two indicators:

Polarity, which ranges from -1 (most negative) to 1 (most positive). Neutral text gets polarity of 0 or near it;

Subjectivity, which ranges from 0 (not subjective) to 1 (very subjective);

vaderSentiment calculates four indicators:

Compound is like TextBlob's polarity;

Indicators of the proportion of positivity, negativity and neutrality in the text. The three indicators sum to 1;

This function is an implementation of Python function sentiment_scores in package pxtextmining.